Complete basic bloom filter
Add bloom filter in find option.

When use find option,Bloom will process first.
If the word isn't found by Bloom filter,we can
say that word isn't in the list.
But if Bloom filter fount it.We can only claim
that word "maybe" in the tree,So I make it to
walk through the tree.

By the way,the new file "ErrorReport" in unit-test
is used to Reoprt bug in program